requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
docker>=6.0.0
